refactor ui/tiles: tiles/type + tiles/util + renaming/merging tiles/gutters into tiles/sidebar
